nuamproject/urls.py
from django.contrib import admin
from django.urls import path, include
from drf_spectacular.views import (
    SpectacularAPIView,
    SpectacularSwaggerView
)

urlpatterns = [
    # Admin Django
    path("admin/", admin.site.urls),
    # FRONTEND (HTML)
    path("", include("frontend.urls")), 
    # API (REST)
    path("api/", include("api.urls")),   
    # OpenAPI Schema
    path("schema/", SpectacularAPIView.as_view(), name="schema"),
    # Swagger UI
    path(
        "docs/",
        SpectacularSwaggerView.as_view(url_name="schema"),
        name="swagger-ui",
    ),
]
